requirements.txt
pycryptodomex
pillow
pyngrok